Improve imports of data structures
