+++ openscap_report/scap_results_parser/data_structures/report.py
@@ -2,7 +2,6 @@ import logging
 from dataclasses import dataclass
 
 from ..exceptions import MissingProcessableRules
-from .oval_data_structures import OvalNode
 
 
 @dataclass
@@ -98,105 +97,3 @@ class Report:  # pylint: disable=R0902
 
     def get_failed_rules(self):
         return list(filter(lambda rule: rule.result.lower() == "fail", self.rules.values()))
-
-
-@dataclass
-class Rule:  # pylint: disable=R0902
-    rule_id: str = ""
-    title: str = ""
-    result: str = ""
-    multi_check: bool = False
-    time: str = ""
-    severity: str = ""
-    identifiers: list = None
-    references: list = None
-    description: str = ""
-    rationale: str = ""
-    warnings: list = None
-    platforms: list = None
-    oval_definition_id: str = ""
-    messages: list = None
-    remediations: list = None
-    oval_tree: OvalNode = None
-    cpe_tree: OvalNode = None
-
-    def as_dict(self):
-        return {
-            "rule_id": self.rule_id,
-            "title": self.title,
-            "result": self.result,
-            "multi_check": self.multi_check,
-            "time": self.time,
-            "severity": self.severity,
-            "identifiers": self.identifiers,
-            "references": self.references,
-            "description": self.description,
-            "rationale": self.rationale,
-            "warnings": self.warnings,
-            "platforms": self.platforms,
-            "oval_definition_id": self.oval_definition_id,
-            "messages": self.messages,
-            "remediations": self.remediations,
-            "oval_tree": self.oval_tree.as_dict(),
-            "cpe_tree": self.cpe_tree.as_dict(),
-        }
-
-
-@dataclass
-class Remediation:
-    remediation_id: str = ""
-    system: str = ""
-    complexity: str = ""
-    disruption: str = ""
-    strategy: str = ""
-    fix: str = ""
-
-    def as_dict(self):
-        return {
-            "remediation_id": self.remediation_id,
-            "system": self.system,
-            "complexity": self.complexity,
-            "disruption": self.disruption,
-            "strategy": self.strategy,
-            "fix": self.fix,
-        }
-
-    def get_type(self):
-        script_types = {
-            "urn:xccdf:fix:script:sh": "Shell script",
-            "urn:xccdf:fix:script:ansible": "Ansible snippet",
-            "urn:xccdf:fix:script:puppet": "Puppet snippet",
-            "urn:redhat:anaconda:pre": "Anaconda snippet",
-            "urn:xccdf:fix:script:kubernetes": "Kubernetes snippet",
-            "urn:redhat:osbuild:blueprint": "OSBuild Blueprint snippet",
-        }
-        return script_types.get(self.system, "script")
-
-
-@dataclass
-class Group:
-    group_id: str = ""
-    title: str = ""
-    description: str = ""
-    platforms: list = None
-    rules_ids: list = None
-    sub_groups: list = None
-
-    def as_dict(self):
-        if not self.sub_groups:
-            return {
-                "group_id": self.group_id,
-                "title": self.title,
-                "description": self.description,
-                "platforms": self.platforms,
-                "rules_ids": self.rules_ids,
-                "sub_groups": None,
-            }
-        return {
-            "group_id": self.group_id,
-            "title": self.title,
-            "description": self.description,
-            "platforms": self.platforms,
-            "rules_ids": self.rules_ids,
-            "sub_groups": [sub_group.as_dict() for sub_group in self.sub_groups],
-        }
