+++ openscap_report/scap_results_parser/parsers/remediation_parser.py
@@ -1,6 +1,6 @@
 from lxml import etree
 
-from ..data_structures.data_structures import Remediation
+from ..data_structures import Remediation
 
 
 class RemediationParser():
