+++ openscap_report/scap_results_parser/parsers/rule_parser.py
@@ -1,4 +1,4 @@
-from ..data_structures.data_structures import Rule
+from ..data_structures import Rule
 from ..namespaces import NAMESPACES
 from .description_parser import DescriptionParser
 from .remediation_parser import RemediationParser
